feat(restapi): merge with existing validation
